Rename extension to SharedDomain and add register_user callback
