--- lib/lightning/extensions/create_project.ex
@@ -1,8 +1,0 @@
-defmodule Lightning.Extensions.CreateProject do
-  @moduledoc """
-  Initializes a project for Lightning API endpoints.
-  """
-
-  @callback create_project(attrs :: map()) ::
-              {:ok, map()} | {:error, any()}
-end
